Remove flash operations from FlashRegion trait and move to common module
